refactor de nombre
