+++ routers/moviesRoute.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from controllers.movies_controllers import movie_router
+from controllers.moviesController import movie_router
 
 routerMovie = APIRouter()
 routerMovie.include_router(movie_router)
